+++ Packages/com.cdm.figma.ui/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.4.12] - 2023-06-14
+### Fixes
+- Fixed svg generation if a node has more than one geometry
+- Fixed svg stroke generation
+
 ## [1.4.10] - 2023-06-11
 ### Fixes
 - Added `Query` extension to `FigmaNode` method searching for component type without binding key